Projects/project 3 select/issue 22 select area

+++ api/SharePaint.Models/Coord2D.cs
@@ -2,8 +2,7 @@
 {
     public class Coord2D
     {
-        // TODO: Decide whether to use int or double in the model
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
     }
 }
